+++ library/src/main/java/com/github/javiersantos/materialstyleddialogs/utils/UtilsAnimation.java
@@ -1,13 +1,13 @@
-package com.github.javiersantos.materialstyleddialogs;
+package com.github.javiersantos.materialstyleddialogs.utils;
 
 import android.content.Context;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 
-class UtilsAnimation {
+public class UtilsAnimation {
 
-    static void zoomInAndOutAnimation(Context context, final ImageView image) {
+    public static void zoomInAndOutAnimation(Context context, final ImageView image) {
         Animation zoomInAnimation = AnimationUtils.loadAnimation(context, R.anim.md_styled_zoom_in);
         final Animation zoomOutAnimation = AnimationUtils.loadAnimation(context, R.anim.md_styled_zoom_out);
